abstract out a couple components
